Moved OrthoSolver to solvers.hpp/cpp.
