+++ miniapps/navier/CMakeLists.txt
@@ -11,11 +11,9 @@
 
 if (MFEM_USE_MPI)
   list(APPEND NAVIER_COMMON_SOURCES
-      navier_solver.cpp
-      ortho_solver.cpp)
+      navier_solver.cpp)
   list(APPEND NAVIER_COMMON_HEADERS
-      navier_solver.hpp
-      ortho_solver.hpp)
+      navier_solver.hpp)
 
   convert_filenames_to_full_paths(NAVIER_COMMON_SOURCES)
   convert_filenames_to_full_paths(NAVIER_COMMON_HEADERS)
