+++ miniapps/navier/navier_solver.hpp
@@ -15,7 +15,6 @@
 #define NAVIER_VERSION 0.1
 
 #include "mfem.hpp"
-#include "ortho_solver.hpp"
 
 namespace mfem
 {
